requirements.txt
pandas        
numpy       
seaborn
statsmodels.stats.api
matplotlib.pyplot
plotly.express
streamlit   
plotly.graph_objects